Manager.java methods all functional; TODO: modify changeReservation to either a. accept more arguments for attribute updates or b. direct user to handle attribute updates

+++ src/main/java/com/umgc/swen646/Reservation.java
@@ -230,7 +230,7 @@ public abstract class Reservation {
         this.bathrooms = bathrooms;
     }
 
-    public void setSquareFootage(int squareFootage) {
+    public void setSquareFootage(double squareFootage) {
         this.squareFootage = squareFootage;
     }
